requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
filelock==3.8.0
flower==1.2.0
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.6
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.1
